+++ R/5-classifier-model-functions.R
@@ -9,11 +9,15 @@
 #'                      returned
 #' @param log logical, if \code{TRUE} log-probabilities are returned
 #' @param ... additional parameters, see details
+#' @details Predict the most probable a posteriori value for the class variable
+#' given all the other variables in the model. Ties are broken at random and 
+#'  if, for a given vector of predictor variables, all conditional probabilities
+#'  are 0, NA is returned.
 #' @return A vector of predictions or the corresponding probabilities
 #' @examples
 #' DD <- generate_xor_dataset(n = 5, 1000)
 #' order <- c("C", "X1", "X2", "X3", "X4", "X5")
-#' train <- DD[1:500, order]
+#' train <- DD[1:5, order]
 #' test <- DD[501:1000, order]
 #' model <- full(train)
 #' model <- bhc.sevt(model)
@@ -89,8 +93,6 @@ predict.sevt <-
       }
     } else{
       class_values <- colnames(pred)
-      return(apply(pred, MARGIN = 1, function(x) {
-        factor(class_values[which.max(x)], levels = class_values)
-      }))
+      return(apply(pred, MARGIN = 1, which_class, levels = class_values))
     }
   }